GT-2894 - Key Bindings - review fixes
