+++ Ghidra/Features/Base/src/main/java/ghidra/app/plugin/core/compositeeditor/CompositeEditorProvider.java
@@ -104,26 +104,26 @@ public abstract class CompositeEditorProvider extends ComponentProviderAdapter
 	}
 
 	protected void addActionsToTool() {
-		CompositeEditorAction[] allActions = actionMgr.getAllActions();
-		for (CompositeEditorAction allAction : allActions) {
+		CompositeEditorTableAction[] allActions = actionMgr.getAllActions();
+		for (CompositeEditorTableAction allAction : allActions) {
 			tool.addLocalAction(this, allAction);
 		}
 	}
 
-	protected CompositeEditorAction[] getActions() {
+	protected CompositeEditorTableAction[] getActions() {
 		return actionMgr.getAllActions();
 	}
 
 	@Override
-	public void actionsAdded(CompositeEditorAction[] actions) {
-		for (CompositeEditorAction action : actions) {
+	public void actionsAdded(CompositeEditorTableAction[] actions) {
+		for (CompositeEditorTableAction action : actions) {
 			tool.addLocalAction(this, action);
 		}
 	}
 
 	@Override
-	public void actionsRemoved(CompositeEditorAction[] actions) {
-		for (CompositeEditorAction action : actions) {
+	public void actionsRemoved(CompositeEditorTableAction[] actions) {
+		for (CompositeEditorTableAction action : actions) {
 			tool.removeLocalAction(this, action);
 		}
 	}
@@ -206,8 +206,8 @@ public abstract class CompositeEditorProvider extends ComponentProviderAdapter
 
 	@Override
 	public void dispose() {
-		CompositeEditorAction[] allActions = actionMgr.getAllActions();
-		for (CompositeEditorAction allAction : allActions) {
+		CompositeEditorTableAction[] allActions = actionMgr.getAllActions();
+		for (CompositeEditorTableAction allAction : allActions) {
 			tool.removeLocalAction(this, allAction);
 		}
 		tool.showComponentProvider(this, false);
@@ -254,8 +254,8 @@ public abstract class CompositeEditorProvider extends ComponentProviderAdapter
 		tool.setStatusInfo(msg);
 	}
 
-	protected CompositeEditorAction[] createActions() {
-		return new CompositeEditorAction[0];
+	protected CompositeEditorTableAction[] createActions() {
+		return new CompositeEditorTableAction[0];
 	}
 
 	protected boolean applyChanges() {
