+++ Ghidra/Features/Base/src/test.slow/java/ghidra/app/plugin/core/compositeeditor/AbstractStructureEditorTest.java
@@ -111,7 +111,7 @@ public abstract class AbstractStructureEditorTest extends AbstractEditorTest {
 
 	void getActions() {
 		actions = provider.getActions();
-		for (CompositeEditorAction action : actions) {
+		for (CompositeEditorTableAction action : actions) {
 			if (action instanceof FavoritesAction) {
 				favorites.add((FavoritesAction) action);
 			}
