+++ Ghidra/Features/Base/src/test.slow/java/ghidra/app/plugin/core/compositeeditor/UnionEditorAlignmentTest.java
@@ -111,7 +111,7 @@ public class UnionEditorAlignmentTest extends AbstractUnionEditorTest {
 		addDataType(arrayDt);
 
 		// Check enablement.
-		CompositeEditorAction[] pActions = provider.getActions();
+		CompositeEditorTableAction[] pActions = provider.getActions();
 		for (int i = 0; i < pActions.length; i++) {
 			if ((pActions[i] instanceof FavoritesAction) ||
 				(pActions[i] instanceof CycleGroupAction) ||
